[REFACTOR(Card)] 카드 뽑기 테스트 추가
